refactor: create modules
