+++ src/js/constants/keyboardConstants.js
@@ -1,4 +1,7 @@
 export default {
+
+    isEnglishLayout: '',
+
     shiftKeyboardLayout: {
         RU: ['Ё', '!', '"', '№', ';', '%', ':', '?', '*', '(',
             ')', '_', '+', '{', '}', ':', '"', '<', '>', '?'],
